+++ COW/src/com/game/Game.java
@@ -132,7 +132,7 @@ public abstract class Game implements GameInterface {
 	 * {@inheritDoc}
 	 */
 	@Override
-	public void setScore(short aiId, long score) {
+	public void setScore(short aiId, int score) {
 		simulator.setScore(aiId, score);
 	}
 	
@@ -140,7 +140,7 @@ public abstract class Game implements GameInterface {
 	 * {@inheritDoc}
 	 */
 	@Override
-	public void incrementScore(short aiId, long increment) {
+	public void incrementScore(short aiId, int increment) {
 		simulator.incrementScore(aiId, increment);
 	}
